refactor(health): merge health endpoint tests into test_health_check_filter
